backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
openai>=1.6.1,<2.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
pypdf==3.17.4
python-docx==1.1.0
chromadb==0.4.22
sentence-transformers==2.2.2
numpy==1.26.2
pandas==2.1.4
aiofiles==23.2.1
httpx==0.25.2
PyPDF2==3.0.1
